Split agent-detection into title status/display/identity domain modules (title evidence plan, PR 3)
Phase 3 of the terminal title evidence plan: the 501-line
shared/agent-detection.ts becomes a 34-line compatibility barrel over
terminal-title-status.ts (classification + stateful tracker),
terminal-title-display.ts (normalization/clearing), and the existing
terminal-title-agent-type.ts (identity predicates, glyphs, getAgentLabel).
Every moved function body is byte-for-byte verbatim; zero consumer or
test edits; dead AGENT_NAMES re-export dropped.
